Styling updates, update track functionality to make track image nullable and use placeholder img or profile img when not provided

+++ localsound.backend/Infrastructure/Repositories/TrackRepository.cs
@@ -159,7 +159,12 @@ namespace localsound.backend.Infrastructure.Repositories
                 if (newTrackImage != null)
                 {
                     await _dbContext.FileContent.AddAsync(newTrackImage);
-                    _dbContext.FileContent.Remove(track.TrackImage);
+
+                    if (track.TrackImage != null)
+                    {
+                        _dbContext.FileContent.Remove(track.TrackImage);
+                    }
+
                     track.TrackImage = newTrackImage;
                     track.TrackImageUrl = newTrackImageUrl;
                 }
